remove setup_hook_httpntlmauth from built-in hooks
